# docs/05-api-openapi.yaml
openapi: 3.0.3
info:
  title: CareerAI API
  description: |
    API REST de CareerAI (Next.js App Router).
    Authentification : session Supabase via cookies (navigateur) ou cookie de session après login.
    Base URL production : https://www.careerai.live
  version: 1.0.0
  contact:
    name: CareerAI
    url: https://www.careerai.live
servers:
  - url: https://www.careerai.live
    description: Production
  - url: http://localhost:3000
    description: Développement local
...
paths:
...
  /api/recruiter/candidates:
    get:
      tags: [recruiter]
      summary: Liste des candidats du recruteur
      security:
        - cookieAuth: []
      parameters:
        - name: jobPostingId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste candidats
          content:
            application/json:
              schema:
                type: object
                properties:
                  candidates:
                    type: array
                    items:
                      $ref: '#/components/schemas/Candidate'
    post:
      tags: [recruiter]
      summary: Ajouter un candidat et analyser son CV
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [email, cv]
              properties:
                email:
                  type: string
                  format: email
                firstName:
                  type: string
                lastName:
                  type: string
                jobPostingId:
                  type: string
                  format: uuid
                cv:
                  type: string
                  format: binary
      responses:
        '200':
          description: Candidat créé
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Email déjà utilisé pour ce recruteur
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Session Supabase (cookies après login). Voir documentation Supabase Auth.
  responses:
    Unauthorized:
      description: Non autorisé — session absente ou expirée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Candidate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        status:
          type: string